refactor uAMQP and Rust AMQP into separate implementations for ease of use
