+++ sdk/core/azure-core-amqp/samples/internal/eventhub_token_writer_sample/eventhub_token_writer_sample.cpp
@@ -73,12 +73,21 @@ int main()
   while (messageSendCount < maxMessageSendCount)
   {
     auto result = sender.Send(message);
+#if ENABLE_UAMQP
     if (std::get<0>(result) != Azure::Core::Amqp::_internal::MessageSendStatus::Ok)
     {
       std::cout << "Failed to send message " << messageSendCount << std::endl;
       std::cout << "Message error information: " << std::get<1>(result) << std::endl;
       break;
     }
+#elif ENABLE_RUST_AMQP
+    if (result)
+    {
+      std::cout << "Failed to send message " << messageSendCount << std::endl;
+      std::cout << "Message error information: " << result << std::endl;
+      break;
+    }
+#endif
     messageSendCount += 1;
   }
