+++ sdk/core/azure-core-amqp/src/amqp/connection_string_credential.cpp
@@ -124,17 +124,6 @@ namespace Azure { namespace Core { namespace Amqp { namespace _internal {
     }
   }
 
-#if ENABLE_UAMQP
-  std::shared_ptr<Network::_internal::Transport>
-  ServiceBusSasConnectionStringCredential::GetTransport() const
-  {
-    //    // Construct a SASL Anonymous transport
-    //    return std::make_shared<Network::SaslTransport>(GetHostName(), GetPort());
-    return std::make_shared<Network::_internal::Transport>(
-        Network::_internal::SocketTransportFactory::Create(GetHostName(), GetPort()));
-  }
-#endif
-
   Credentials::AccessToken ServiceBusSasConnectionStringCredential::GetToken(
       Credentials::TokenRequestContext const& tokenRequestContext,
       Context const& context) const
@@ -147,57 +136,4 @@ namespace Azure { namespace Core { namespace Amqp { namespace _internal {
     return rv;
   }
 
-  // Generate a Shared Access Signature token for a ServiceBus client.
-  //
-  // The spec for a SharedAccessSignature is here:
-  // https://learn.microsoft.com/en-us/azure/service-bus-messaging/service-bus-sas#generate-a-shared-access-signature-token
-  // Samples for SAS generation are here:
-  // https://learn.microsoft.com/en-us/rest/api/eventhub/generate-sas-token
-  //
-  std::string ServiceBusSasConnectionStringCredential::GenerateSasToken(
-      std::chrono::system_clock::time_point const& expirationTime) const
-  {
-#if ENABLE_UAMQP
-    // For now, create the SAS token using the azure-c-shared-utility functions for SAS token
-    // creations. In the future (when we've integrated into Azure Core), move this to the Azure Core
-    // APIs.
-
-    /* Construct a SAS token */
-    /* Base64 encode the SharedAccessKey - this needs to be double encoded because of <reasons> */
-    std::vector<uint8_t> dataToEncode;
-    dataToEncode.reserve(GetSharedAccessKey().size());
-    std::transform(
-        std::begin(GetSharedAccessKey()),
-        std::end(GetSharedAccessKey()),
-        std::back_inserter(dataToEncode),
-        [](char ch) { return static_cast<uint8_t>(ch); });
-
-    std::string encodedKeyValue{Azure::Core::Convert::Base64Encode(dataToEncode)};
-    Azure::Core::Url resourceUri{GetEndpoint()};
-    resourceUri.AppendPath(GetEntityPath());
-
-    auto sasKeyValue{STRING_construct(encodedKeyValue.c_str())};
-    auto encodedResourceUri = URL_EncodeString(resourceUri.GetAbsoluteUrl().c_str());
-    auto sasKeyName = STRING_construct(GetSharedAccessKeyName().c_str());
-    auto sasToken = SASToken_Create(
-        sasKeyValue,
-        encodedResourceUri,
-        sasKeyName,
-        std::chrono::duration_cast<std::chrono::seconds>(expirationTime.time_since_epoch())
-            .count());
-    if (sasToken == nullptr)
-    {
-      throw std::runtime_error("Could not create SAS token.");
-    }
-    std::string rv(STRING_c_str(sasToken));
-    STRING_delete(sasToken);
-    STRING_delete(sasKeyValue);
-    STRING_delete(sasKeyName);
-    STRING_delete(encodedResourceUri);
-    return rv;
-#else
-    (void)expirationTime;
-    return std::string();
-#endif // ENABLE_UAMQP
-  }
 }}}} // namespace Azure::Core::Amqp::_internal
