+++ sdk/core/azure-core-amqp/src/impl/rust_amqp/amqp/private/connection_impl.hpp
