+++ sdk/core/azure-core-amqp/src/impl/rust_amqp/amqp/private/link_impl.hpp
