+++ sdk/core/azure-core-amqp/src/impl/rust_amqp/amqp/private/message_sender_impl.hpp
@@ -1,0 +1,77 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+#pragma once
+
+#include "azure/core/amqp/internal/message_sender.hpp"
+#include "link_impl.hpp"
+#include "unique_handle.hpp"
+
+namespace Azure { namespace Core { namespace Amqp { namespace _detail {
+#if ENABLE_UAMQP
+  template <> struct UniqueHandleHelper<MESSAGE_SENDER_INSTANCE_TAG>
+  {
+    static void FreeMessageSender(MESSAGE_SENDER_HANDLE obj);
+
+    using type = Core::_internal::BasicUniqueHandle<MESSAGE_SENDER_INSTANCE_TAG, FreeMessageSender>;
+  };
+#endif
+}}}} // namespace Azure::Core::Amqp::_detail
+
+namespace Azure { namespace Core { namespace Amqp { namespace _detail {
+#if ENABLE_UAMQP
+  using UniqueMessageSender = UniqueHandle<MESSAGE_SENDER_INSTANCE_TAG>;
+#endif
+
+  class MessageSenderFactory final {
+  public:
+    static Azure::Core::Amqp::_internal::MessageSender CreateFromInternal(
+        std::shared_ptr<MessageSenderImpl> senderImpl)
+    {
+      return Azure::Core::Amqp::_internal::MessageSender(senderImpl);
+    }
+  };
+
+  class MessageSenderImpl : public std::enable_shared_from_this<MessageSenderImpl> {
+  public:
+    MessageSenderImpl(
+        std::shared_ptr<_detail::SessionImpl> session,
+        Models::_internal::MessageTarget const& target,
+        _internal::MessageSenderOptions const& options);
+    virtual ~MessageSenderImpl() noexcept;
+
+    MessageSenderImpl(MessageSenderImpl const&) = delete;
+    MessageSenderImpl& operator=(MessageSenderImpl const&) = delete;
+    MessageSenderImpl(MessageSenderImpl&&) noexcept = delete;
+    MessageSenderImpl& operator=(MessageSenderImpl&&) noexcept = delete;
+
+    Models::_internal::AmqpError Open(bool blockingOpen, Context const& context);
+    void Close(Context const& context);
+    Models::_internal::AmqpError Send(Models::AmqpMessage const& message, Context const& context);
+
+    std::uint64_t GetMaxMessageSize() const;
+
+    std::string GetLinkName() const;
+
+  private:
+    void CreateLink();
+    void PopulateLinkProperties();
+    void OnLinkDetached(Models::_internal::AmqpError const& error);
+
+    bool m_senderOpen{false};
+#if ENABLE_UAMQP
+    UniqueMessageSender m_messageSender{};
+#endif
+    std::shared_ptr<_detail::LinkImpl> m_link;
+    Models::_internal::AmqpError m_savedMessageError;
+
+    Azure::Core::Amqp::Common::_internal::AsyncOperationQueue<Models::_internal::AmqpError>
+        m_openQueue;
+    Azure::Core::Amqp::Common::_internal::AsyncOperationQueue<Models::_internal::AmqpError>
+        m_closeQueue;
+
+    std::shared_ptr<_detail::SessionImpl> m_session;
+    Models::_internal::MessageTarget m_target;
+    _internal::MessageSenderOptions m_options;
+  };
+}}}} // namespace Azure::Core::Amqp::_detail
