+++ sdk/core/azure-core-amqp/src/impl/rust_amqp/amqp/private/session_impl.hpp
