+++ sdk/core/azure-core-amqp/src/impl/uamqp/amqp/private/management_impl.hpp
