+++ sdk/core/azure-core-amqp/src/impl/uamqp/network/private/transport_impl.hpp
@@ -4,7 +4,7 @@
 #pragma once
 #if ENABLE_UAMQP
 
-#include "../../amqp/private/unique_handle.hpp"
+#include "../../../amqp/private/unique_handle.hpp"
 #include "azure/core/amqp/internal/common/async_operation_queue.hpp"
 
 #include <azure_c_shared_utility/xio.h>
