+++ sdk/core/azure-core-amqp/src/impl/uamqp/network/sasl_transport.cpp
