+++ sdk/core/azure-core-amqp/src/impl/uamqp/network/socket_transport.cpp
